Extract graph_recurrence
